feat: date and time section
